% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_all_Model_ID}
\alias{dd_dat_all_Model_ID}
\title{Title}
\usage{
dd_dat_all_Model_ID(
  list_dat_features = list(dat_client_GER_Model_Date_features,
    dat_client_Syncronys_Model_Date_features,
    dat_client_Conduent_Omnicad_Model_Date_features, dat_client_RORA_Model_Date_features,
    dat_client_BBS_Model_Date_features, dat_client_CaseNotes),
  by_for_join = dplyr::join_by(ANE_Substantiated, Client_System_ID, Client_Gender,
    Client_Ethnicity, Client_Race, Client_Region, Age)
)
}
\arguments{
\item{list_dat_features}{list_dat_features}

\item{by_for_join}{by_for_join}
}
\value{
dat_all_Model_ID
}
\description{
Title
}
