% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_BBS_Model}
\alias{dd_dat_client_BBS_Model}
\title{Title}
\usage{
dd_dat_client_BBS_Model(dat_client_BBS = dat_client_BBS)
}
\arguments{
\item{dat_client_BBS}{dat_client_BBS}
}
\value{
dat_client_BBS_Model
}
\description{
Title
}
