% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_BBS_Model_Date}
\alias{dd_dat_client_BBS_Model_Date}
\title{Title}
\usage{
dd_dat_client_BBS_Model_Date(
  dat_client_BBS_Model = dat_client_BBS_Model,
  dat_client_IMB_ANE_Model = dat_client_IMB_ANE_Model,
  date_Current = date_Current,
  sw_ANE_Current = c("ANE", "Current")[1]
)
}
\arguments{
\item{dat_client_BBS_Model}{dat_client_BBS_Model}

\item{dat_client_IMB_ANE_Model}{dat_client_IMB_ANE_Model}

\item{date_Current}{date_Current}

\item{sw_ANE_Current}{sw_ANE_Current}
}
\value{
dat_client_BBS_Model_Date
}
\description{
Title
}
