% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_CaseNotes_Model}
\alias{dd_dat_client_CaseNotes_Model}
\title{Title}
\usage{
dd_dat_client_CaseNotes_Model(dat_client_CaseNotes = dat_client_CaseNotes)
}
\arguments{
\item{dat_client_CaseNotes}{dat_client_CaseNotes}
}
\value{
dat_client_CaseNotes_Model
}
\description{
Title
}
