% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_CaseNotes_Model_Date_features}
\alias{dd_dat_client_CaseNotes_Model_Date_features}
\title{Title}
\usage{
dd_dat_client_CaseNotes_Model_Date_features(
  dat_client_Match_Model = dat_client_Match_Model,
  dat_client_IMB_ANE_Model = dat_client_IMB_ANE_Model,
  dat_client_CaseNotes_Model = dat_client_CaseNotes_Model,
  date_Current = date_Current,
  sw_ANE_Current = c("ANE", "Current")[1],
  m_months_CaseNotes = m_months_CaseNotes
)
}
\arguments{
\item{dat_client_Match_Model}{dat_client_Match_Model}

\item{dat_client_IMB_ANE_Model}{dat_client_IMB_ANE_Model}

\item{dat_client_CaseNotes_Model}{dat_client_CaseNotes_Model}

\item{date_Current}{date_Current}

\item{sw_ANE_Current}{sw_ANE_Current}
}
\value{
dat_client_CaseNotes_Model_Date_features
}
\description{
Title
}
