% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_eCHAT_Model}
\alias{dd_dat_client_eCHAT_Model}
\title{Title}
\usage{
dd_dat_client_eCHAT_Model(dat_client_eCHAT = dat_client_eCHAT)
}
\arguments{
\item{dat_client_eCHAT}{dat_client_eCHAT}
}
\value{
dat_client_eCHAT_Model
}
\description{
Title
}
