% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_eCHAT_Model_Date_features_Model}
\alias{dd_dat_client_eCHAT_Model_Date_features_Model}
\title{Title}
\usage{
dd_dat_client_eCHAT_Model_Date_features_Model(
  dat_client_Match_Model = dat_client_Match_Model,
  dat_client_IMB_ANE_Model = dat_client_IMB_ANE_Model,
  dat_client_eCHAT_Model_Date = dat_client_eCHAT_Model_Date,
  date_Current = date_Current,
  sw_ANE_Current = c("ANE", "Current")[1],
  m_months_eCHAT = NULL
)
}
\arguments{
\item{dat_client_Match_Model}{dat_client_Match_Model}

\item{dat_client_IMB_ANE_Model}{dat_client_IMB_ANE_Model}

\item{dat_client_eCHAT_Model_Date}{dat_client_eCHAT_Model_Date}

\item{date_Current}{date_Current}

\item{sw_ANE_Current}{sw_ANE_Current}

\item{m_months_eCHAT}{m_months_eCHAT}
}
\value{
dat_client_eCHAT_Model_Date_features_Model
}
\description{
Title
}
