% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_GER_Model}
\alias{dd_dat_client_GER_Model}
\title{Title}
\usage{
dd_dat_client_GER_Model(dat_client_GER = dat_client_GER)
}
\arguments{
\item{dat_client_GER}{dat_client_GER}
}
\value{
dat_client_GER_Model
}
\description{
Title
}
