% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_IMB_ANE_Model}
\alias{dd_dat_client_IMB_ANE_Model}
\title{Title}
\usage{
dd_dat_client_IMB_ANE_Model(dat_client_IMB_ANE = dat_client_IMB_ANE)
}
\arguments{
\item{dat_client_IMB_ANE}{dat_client_IMB_ANE}
}
\value{
dat_client_IMB_ANE_Model
}
\description{
Title
}
