% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_Match_Model}
\alias{dd_dat_client_Match_Model}
\title{Title}
\usage{
dd_dat_client_Match_Model(dat_client_Match = dat_client_Match)
}
\arguments{
\item{dat_client_Match}{dat_client_Match}
}
\value{
dat_client_Match_Model
}
\description{
Title
}
