% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_RORA_Model}
\alias{dd_dat_client_RORA_Model}
\title{Title}
\usage{
dd_dat_client_RORA_Model(dat_client_RORA = dat_client_RORA)
}
\arguments{
\item{dat_client_RORA}{dat_client_RORA}
}
\value{
dat_client_RORA_Model
}
\description{
Title
}
