% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_dat_client_Syncronys_Model}
\alias{dd_dat_client_Syncronys_Model}
\title{Title}
\usage{
dd_dat_client_Syncronys_Model(dat_client_Syncronys = dat_client_Syncronys)
}
\arguments{
\item{dat_client_Syncronys}{dat_client_Syncronys}
}
\value{
dat_client_Syncronys_Model
}
\description{
Title
}
