% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_BBS.R
\name{dd_features_client_BBS}
\alias{dd_features_client_BBS}
\title{Features for Client BBS}
\usage{
dd_features_client_BBS(dat_client_BBS = NULL)
}
\arguments{
\item{dat_client_BBS}{dat_client_BBS data from \code{dd_read_client_BBS()}}
}
\value{
dat_client_BBS
}
\description{
Features for Client BBS
}
\examples{
\dontrun{
dd_features_client_BBS(
    dat_client_BBS
  )
}
}
