% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_CaseNotes.R
\name{dd_features_client_CaseNotes}
\alias{dd_features_client_CaseNotes}
\title{Features for Client CaseNotes}
\usage{
dd_features_client_CaseNotes(
  dat_client_CaseNotes = NULL,
  dat_client_Match = dat_client_Match
)
}
\arguments{
\item{dat_client_CaseNotes}{dat_client_CaseNotes data from \code{dd_read_client_CaseNotes()}}

\item{dat_client_Match}{dat_client_Match data from \code{dd_read_client_Match()}}
}
\value{
dat_client_CaseNotes
}
\description{
Features for Client CaseNotes
}
\examples{
\dontrun{
dd_features_client_CaseNotes(
    dat_client_CaseNotes
  )
}
}
