% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_Conduent_Omnicad.R
\name{dd_features_client_Conduent_Omnicad}
\alias{dd_features_client_Conduent_Omnicad}
\title{Features for Client Conduent Omnicad}
\usage{
dd_features_client_Conduent_Omnicad(
  dat_client_Conduent_Omnicad = NULL,
  dat_client_Match = dat_client_Match
)
}
\arguments{
\item{dat_client_Conduent_Omnicad}{dat_client_Conduent_Omnicad data from \code{dd_read_client_Conduent_Omnicad()}}

\item{dat_client_Match}{dat_client_Match data from \code{dd_read_client_Match()}}
}
\value{
dat_client_Conduent_Omnicad
}
\description{
Features for Client Conduent Omnicad
}
\examples{
\dontrun{
dd_features_client_Conduent_Omnicad(
    dat_client_Conduent_Omnicad
  )
}
}
