% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_Conduent_Omnicad_read.R
\name{dd_features_client_Conduent_Omnicad_read}
\alias{dd_features_client_Conduent_Omnicad_read}
\title{Features for Client Conduent Omnicad during read}
\usage{
dd_features_client_Conduent_Omnicad_read(dat_client_Conduent_Omnicad = NULL)
}
\arguments{
\item{dat_client_Conduent_Omnicad}{dat_client_Conduent_Omnicad data from \code{dd_read_client_Conduent_Omnicad()}}
}
\value{
dat_client_Conduent_Omnicad
}
\description{
Features for Client Conduent Omnicad during read
}
\examples{
\dontrun{
dd_features_client_Conduent_Omnicad_read(
    dat_client_Conduent_Omnicad
  )
}
}
