% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_eCHAT.R
\name{dd_features_client_eCHAT}
\alias{dd_features_client_eCHAT}
\title{Features for Client eCHAT}
\usage{
dd_features_client_eCHAT(
  dat_client_eCHAT = NULL,
  dat_client_Match = dat_client_Match
)
}
\arguments{
\item{dat_client_eCHAT}{dat_client_eCHAT data from \code{dd_read_client_eCHAT()}}

\item{dat_client_Match}{dat_client_Match data from \code{dd_read_client_Match()}}
}
\value{
dat_client_eCHAT
}
\description{
Features for Client eCHAT
}
\examples{
\dontrun{
dd_features_client_eCHAT(
    dat_client_eCHAT
  , dat_client_Match  = dat_client_Match
  )
}
}
