% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_IMB_ANE.R
\name{dd_features_client_IMB_ANE}
\alias{dd_features_client_IMB_ANE}
\title{Features for Client IMB_ANE}
\usage{
dd_features_client_IMB_ANE(
  dat_client_IMB_ANE = NULL,
  dat_client_Match = dat_client_Match
)
}
\arguments{
\item{dat_client_IMB_ANE}{dat_client_IMB_ANE data from \code{dd_read_client_IMB_ANE()}}

\item{dat_client_Match}{dat_client_Match data from \code{dd_read_client_Match()}}
}
\value{
dat_client_IMB_ANE
}
\description{
Features for Client IMB_ANE
}
\examples{
\dontrun{
dd_features_client_IMB_ANE(
    dat_client_IMB_ANE
  )
}
}
