% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_Match.R
\name{dd_features_client_Match}
\alias{dd_features_client_Match}
\title{Features for Client Match}
\usage{
dd_features_client_Match(dat_client_Match = NULL)
}
\arguments{
\item{dat_client_Match}{dat_client_Match data from \code{dd_read_client_Match()}}
}
\value{
dat_client_Match
}
\description{
Features for Client Match
}
\examples{
\dontrun{
dd_features_client_Match(
    dat_client_Match
  )
}
}
