% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_RORA.R
\name{dd_features_client_RORA}
\alias{dd_features_client_RORA}
\title{Features for Client RORA}
\usage{
dd_features_client_RORA(dat_client_RORA = NULL)
}
\arguments{
\item{dat_client_RORA}{dat_client_RORA data from \code{dd_read_client_RORA()}}
}
\value{
dat_client_RORA
}
\description{
Features for Client RORA
}
\examples{
\dontrun{
dd_features_client_RORA(
    dat_client_RORA
  )
}
}
