% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_Synchonys.R
\name{dd_features_client_Synchonys}
\alias{dd_features_client_Synchonys}
\title{Features for Client Synchonys}
\usage{
dd_features_client_Synchonys(dat_client_Synchonys = NULL)
}
\arguments{
\item{dat_client_Synchonys}{dat_client_Synchonys data from \code{dd_read_client_Synchonys()}}
}
\value{
dat_client_Synchonys
}
\description{
Features for Client Synchonys
}
\examples{
\dontrun{
dd_features_client_Synchonys(
    dat_client_Synchonys
  )
}
}
