% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_Syncronys.R
\name{dd_features_client_Syncronys}
\alias{dd_features_client_Syncronys}
\title{Features for Client Syncronys}
\usage{
dd_features_client_Syncronys(
  dat_client_Syncronys = NULL,
  dat_client_Match = dat_client_Match
)
}
\arguments{
\item{dat_client_Syncronys}{dat_client_Syncronys data from \code{dd_read_client_Syncronys()}}

\item{dat_client_Match}{dat_client_Match data from \code{dd_read_client_Match()}}
}
\value{
dat_client_Syncronys
}
\description{
Features for Client Syncronys
}
\examples{
\dontrun{
dd_features_client_Syncronys(
    dat_client_Syncronys
  )
}
}
