% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_provider_Match.R
\name{dd_features_provider_Match}
\alias{dd_features_provider_Match}
\title{Features for Provider Match}
\usage{
dd_features_provider_Match(dat_provider_Match = NULL)
}
\arguments{
\item{dat_provider_Match}{dat_provider_Match data from \code{dd_read_provider_Match()}}
}
\value{
dat_provider_Match
}
\description{
Features for Provider Match
}
\examples{
\dontrun{
dd_features_provider_Match(
    dat_provider_Match
  )
}
}
