% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_provider_PDS.R
\name{dd_features_provider_PDS}
\alias{dd_features_provider_PDS}
\title{Features for Provider PDS}
\usage{
dd_features_provider_PDS(dat_provider_PDS = NULL)
}
\arguments{
\item{dat_provider_PDS}{dat_provider_PDS data from \code{dd_read_provider_PDS()}}
}
\value{
dat_provider_PDS
}
\description{
Features for Provider PDS
}
\examples{
\dontrun{
dd_features_provider_PDS(
    dat_provider_PDS
  )
}
}
