% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_provider_RORA.R
\name{dd_features_provider_RORA}
\alias{dd_features_provider_RORA}
\title{Features for Provider RORA}
\usage{
dd_features_provider_RORA(dat_provider_RORA = NULL)
}
\arguments{
\item{dat_provider_RORA}{dat_provider_RORA data from \code{dd_read_provider_RORA()}}
}
\value{
dat_provider_RORA
}
\description{
Features for Provider RORA
}
\examples{
\dontrun{
dd_features_provider_RORA(
    dat_provider_RORA
  )
}
}
