% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_is_error.R
\name{dd_is_error}
\alias{dd_is_error}
\title{check a function and if there's an error return an indication that we can ifelse() to either run for real or skip}
\usage{
dd_is_error(f)
}
\arguments{
\item{f}{report result of a \code{try()}}
}
\value{
TRUE/FALSE for error
}
\description{
check a function and if there's an error return an indication that we can ifelse() to either run for real or skip
}
\examples{
sw_try_ok <-
  !dd_is_error(
    try(
      # try what you want, check for error
      log(NULL)
    )
  )
if(sw_try_ok) {
  # no error, do what you intended to do
} else {
  # error, do something else
}
}
