% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assemble_data.R
\name{dd_list_dat_each_Model_Date_features}
\alias{dd_list_dat_each_Model_Date_features}
\title{Title}
\usage{
dd_list_dat_each_Model_Date_features(
  sw_ANE_Current = c("ANE", "Current")[1],
  dat_client_Match = dat_client_Match,
  dat_client_IMB_ANE = dat_client_IMB_ANE,
  dat_client_GER = dat_client_GER,
  dat_client_Syncronys = dat_client_Syncronys,
  dat_client_RORA = dat_client_RORA,
  dat_client_Conduent_Omnicad = dat_client_Conduent_Omnicad,
  dat_client_BBS = dat_client_BBS,
  date_Current = date_Current,
  m_months_GER = m_months_GER,
  m_months_Syncronys = m_months_Syncronys,
  m_months_Conduent_Omnicad = m_months_Conduent_Omnicad,
  m_months_RORA = m_months_RORA
)
}
\arguments{
\item{sw_ANE_Current}{sw_ANE_Current}

\item{dat_client_Match}{dat_client_Match}

\item{dat_client_IMB_ANE}{dat_client_IMB_ANE}

\item{dat_client_GER}{dat_client_GER}

\item{dat_client_Syncronys}{dat_client_Syncronys}

\item{dat_client_RORA}{dat_client_RORA}

\item{dat_client_Conduent_Omnicad}{dat_client_Conduent_Omnicad}

\item{dat_client_BBS}{dat_client_BBS}

\item{date_Current}{date_Current}

\item{m_months_GER}{m_months_GER}

\item{m_months_Syncronys}{m_months_Syncronys}

\item{m_months_Conduent_Omnicad}{m_months_Conduent_Omnicad}

\item{m_months_RORA}{m_months_RORA}
}
\value{
list_dat_each_Model_Date_features
}
\description{
Title
}
