% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_plot_missing_codebook.R
\name{dd_plot_missing_codebook}
\alias{dd_plot_missing_codebook}
\title{Plot data missing values and create codebook}
\usage{
dd_plot_missing_codebook(
  dat_this = NULL,
  name_dat = "",
  path_results_dat = path_results_dat,
  sw_width = 10,
  sw_height = 10,
  sw_codebook = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{dat_this}{data to plot}

\item{name_dat}{name of dataset, such as \code{dat_XXX}}

\item{path_results_dat}{path to results}

\item{sw_width}{missing plot width in inches}

\item{sw_height}{missing plot height in inches}

\item{sw_codebook}{T/F generate codebook}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{             NULL, invisibly
}\if{html}{\out{</div>}}
}
\description{
Plot data missing values and create codebook
}
\examples{
\dontrun{
dd_plot_missing_codebook(
    dat_this         = dat_GER
  , name_dat         = name_dat
  , path_results_dat = path_results_dat
  #, sw_width         = 10
  #, sw_height        = 10
  , sw_codebook      = sw_codebook
  )
}
}
