% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_prediction_model.R
\name{dd_prediction_model}
\alias{dd_prediction_model}
\title{NM DOH DD Prediction Model, main predict and model function}
\usage{
dd_prediction_model(
  working_directory =
    "D:/Dropbox/StatAcumen/consult/Industry/2023_NM_DOH_DDWaiver/data/Zdrive/Risk_Prediction_Model/Scripts",
  path_prefix_out = "out_V10",
  DD_Group = c("All", "DD", "MV", "MF")[1],
  date_Current = "2023-12-01",
  m_months_GER = "6 months",
  m_months_Syncronys = "2 months",
  m_months_Conduent_Omnicad = "2 months",
  m_months_RORA = "2 months",
  m_months_BBS = "12 months",
  m_months_CaseNotes = "2 months",
  sw_rfsrc_ntree = 500,
  sw_alpha = 0.2,
  sw_select_full = c("select", "full")[1],
  sw_imbalanced_binary = c(FALSE, TRUE)[1],
  sw_m_months_select_quick_full = FALSE,
  sw_m_months_values = list(GER = c(1, 2), Syncronys = c(1, 2), Conduent_Omnicad = c(1),
    RORA = c(1), BBS = c(1), CaseNotes = c(1))
)
}
\arguments{
\item{working_directory}{Working directory}

\item{path_prefix_out}{Path and file prefix for output folders}

\item{DD_Group}{All or specific Waiver group}

\item{date_Current}{Date for last day of data to use for analysis}

\item{m_months_GER}{GER       data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_Syncronys}{Syncronys data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_Conduent_Omnicad}{Omnicad   data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_RORA}{RORA      data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_BBS}{BBS       data, date range to keep, from first ANE or last observation back number of months}

\item{sw_rfsrc_ntree}{Random forest, number of trees}

\item{sw_alpha}{Random forest, model selection alpha level}

\item{sw_select_full}{run RF with model selection, or only fit full model}

\item{sw_imbalanced_binary}{T/F input to \code{e_rfstr_classification()}, to use standard or imbalanced binary classification with \code{rfsrc::imbalanced()}.  Also increases ntree to \code{5 * sw_rfsrc_ntree}.}

\item{sw_m_months_select_quick_full}{T/F to only search for best \code{m_months} arguments.}

\item{sw_m_months_values}{if \code{sw_m_months_select_quick_full == TRUE}, list of number of months for each data source to perform model selection on m_months arguments.  One strategy for search is to start with \code{c(1, 2)} and if the result is 1, then keep 1, otherwise try \code{c(2, 3)}, and so on for 4, 6, 9, and 12 months.  Increment only 4 values at a time for 16 processors.}
}
\value{
dat_client_RORA
}
\description{
NM DOH DD Prediction Model, main predict and model function
}
\examples{
\dontrun{
dd_read_client_RORA(
    fn_list           = c(
                          "FY23 RORA Monitoring Data_20230914.xlsx"
                        )
  , path_data         = "../Data_in/Client_RORA"
  , path_results_dat  = path_results_dat
  , sw_plot_missing   = c(TRUE, FALSE)[1]
  , sw_codebook       = c(TRUE, FALSE)[1]
  )
}
}
