% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_read_ALL_features.R
\name{dd_read_ALL_features}
\alias{dd_read_ALL_features}
\title{NM DOH DD Prediction Model, main predict and model function}
\usage{
dd_read_ALL_features(
  params = params,
  path_list = path_list,
  sw_plot_missing = c(TRUE, FALSE)[1],
  sw_codebook = c(TRUE, FALSE)[2]
)
}
\arguments{
\item{params}{\code{params} object from qmd file, \code{NMDOH_DD_RiskPredictionModel_Data-Model-Summary_yyyymmdd.qmd}}

\item{path_list}{list of paths to data and output from \code{dd_set_paths()}}

\item{sw_plot_missing}{T/F Create plot of missing values (quick)}

\item{sw_codebook}{T/F Create a codebook for each dataset (slow)}
}
\value{
NULL               invisibly
}
\description{
NM DOH DD Prediction Model, main predict and model function
}
\examples{
\dontrun{
dd_read_ALL_features <-
  function(
    working_directory   = "D:/Dropbox/StatAcumen/consult/Industry/2023_NM_DOH_DDWaiver/data/Zdrive/Risk_Prediction_Model/Scripts"
  , params              = params
  , sw_plot_missing     = c(TRUE, FALSE)[1]
  , sw_codebook         = c(TRUE, FALSE)[2]
  )
}
}
