% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_read_client_BBS.R
\name{dd_read_client_BBS}
\alias{dd_read_client_BBS}
\title{Read Client BBS}
\usage{
dd_read_client_BBS(
  fn_list = c("FY23 BBS Monitoring Data_20230914.xlsx"),
  path_data = "../Data_in/Client_BBS",
  path_results_dat = NULL,
  sw_plot_missing = c(TRUE, FALSE)[1],
  sw_codebook = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{fn_list}{list of files to read}

\item{path_data}{path to data}

\item{path_results_dat}{path to write .RData file}

\item{sw_plot_missing}{T/F plot missing values}

\item{sw_codebook}{T/F generage codebook (only runs if \code{sw_plot_missing} is also \code{TRUE}}
}
\value{
dat_client_BBS
}
\description{
Read Client BBS
}
\examples{
\dontrun{
dd_read_client_BBS(
    fn_list           = c(
                          "FY23 BBS Monitoring Data_20230914.xlsx"
                        )
  , path_data         = "../Data_in/Client_BBS"
  , path_results_dat  = path_results_dat
  , sw_plot_missing   = c(TRUE, FALSE)[1]
  , sw_codebook       = c(TRUE, FALSE)[1]
  )
}
}
