% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_read_client_GER.R
\name{dd_read_client_GER}
\alias{dd_read_client_GER}
\title{Read Client GER}
\usage{
dd_read_client_GER(
  fn_list = NULL,
  path_data = "../Data_in/Client_GER",
  path_results_dat = NULL,
  sw_plot_missing = c(TRUE, FALSE)[1],
  sw_codebook = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{fn_list}{NULL, no list of files to read (read all in the path)}

\item{path_data}{path to data}

\item{path_results_dat}{path to write .RData file}

\item{sw_plot_missing}{T/F plot missing values}

\item{sw_codebook}{T/F generage codebook (only runs if \code{sw_plot_missing} is also \code{TRUE}}
}
\value{
dat_client_GER
}
\description{
Read Client GER
}
\examples{
\dontrun{
dd_read_client_GER(
    fn_list           = NULL
  , path_data         = "../Data_in/Client_GER"
  , path_results_dat  = path_results_dat
  , sw_plot_missing   = c(TRUE, FALSE)[1]
  , sw_codebook       = c(TRUE, FALSE)[1]
  )
}
}
