% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_read_client_Syncronys.R
\name{dd_read_client_Syncronys}
\alias{dd_read_client_Syncronys}
\title{Read Client Syncronys}
\usage{
dd_read_client_Syncronys(
  fn_list = NULL,
  path_data = "../Data_in/Client_Syncronys",
  path_results_dat = NULL,
  sw_plot_missing = c(TRUE, FALSE)[1],
  sw_codebook = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{fn_list}{NULL, no list of files to read}

\item{path_data}{path to data}

\item{path_results_dat}{path to write .RData file}

\item{sw_plot_missing}{T/F plot missing values}

\item{sw_codebook}{T/F generage codebook (only runs if \code{sw_plot_missing} is also \code{TRUE}}
}
\value{
dat_client_Syncronys
}
\description{
Read Client Syncronys
}
\examples{
\dontrun{
dd_read_client_Syncronys(
    fn_list           = NULL
  , path_data         = "../Data_in/Client_Syncronys"
  , path_results_dat  = path_results_dat
  , sw_plot_missing   = c(TRUE, FALSE)[1]
  , sw_codebook       = c(TRUE, FALSE)[1]
  )
}
}
