% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_read_provider_PDS.R
\name{dd_read_provider_PDS}
\alias{dd_read_provider_PDS}
\title{Read Provider PDS}
\usage{
dd_read_provider_PDS(
  fn_list = c("Scorecard Jan 23.xlsx", "Scorecard June 23.xlsx",
    "Scorecard March 23.xlsx", "Scorecard Oct 22.xlsx"),
  path_data = "../Data_in/Provider_PDS_Scorecard",
  path_results_dat = NULL,
  sw_plot_missing = c(TRUE, FALSE)[1],
  sw_codebook = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{fn_list}{list of files to read}

\item{path_data}{path to data}

\item{path_results_dat}{path to write .RData file}

\item{sw_plot_missing}{T/F plot missing values}

\item{sw_codebook}{T/F generage codebook (only runs if \code{sw_plot_missing} is also \code{TRUE}}
}
\value{
dat_provider_PDS
}
\description{
Read Provider PDS
}
\examples{
\dontrun{
dd_read_provider_PDS(
    fn_list           = c(
                          "Scorecard Jan 23.xlsx"
                        , "Scorecard June 23.xlsx"
                        , "Scorecard March 23.xlsx"
                        , "Scorecard Oct 22.xlsx"
                        )
  , path_data         = "../Data_in/Provider_PDS_Scorecard"
  , path_results_dat  = path_results_dat
  , sw_plot_missing   = c(TRUE, FALSE)[1]
  , sw_codebook       = c(TRUE, FALSE)[1]
  )
}
}
