% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_save_to_RData.R
\name{dd_save_to_RData}
\alias{dd_save_to_RData}
\title{Save any data object to .RData}
\usage{
dd_save_to_RData(name_dat = NULL)
}
\arguments{
\item{name_dat}{name of data object to save
XXXparam path_results_dat  path to write .RData file}
}
\description{
Save any data object to .RData
}
\examples{
\dontrun{
dd_save_to_RData(
    name_dat          = "dat_client_Match"
  )
}
}
