% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_set_paths.R
\name{dd_set_paths}
\alias{dd_set_paths}
\title{Set paths for data and output}
\usage{
dd_set_paths(
  params = params,
  path_dat_in = "../Data_in",
  path_results = "../Results"
)
}
\arguments{
\item{params}{\code{params} object from qmd file, \code{NMDOH_DD_RiskPredictionModel_Data-Model-Summary_yyyymmdd.qmd}}

\item{path_dat_in}{relative path from \code{params$working_directory} to input data folder, such as \code{"../Data_in"}}

\item{path_results}{relative path from \code{params$working_directory} to results, such as \code{"../Results"}}
}
\value{
path_list          list of paths to data and output
}
\description{
Set paths for data and output
}
