% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nmdohddrisk_packages}
\alias{nmdohddrisk_packages}
\title{List all packages in the nmdohddrisk}
\usage{
nmdohddrisk_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include nmdohddrisk in the list?}
}
\description{
List all packages in the nmdohddrisk
}
\examples{
nmdohddrisk_packages()
}
