% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_xlsx__dd_read_provider_RORA.R
\name{OLD_dd_read_provider_RORA}
\alias{OLD_dd_read_provider_RORA}
\title{Read Provider RORA}
\usage{
OLD_dd_read_provider_RORA(
  fn_list = c("Copy of RORA 09.11.2023 (002).xlsx"),
  path_data = "../Data_in/Provider_RORA",
  path_results_dat = NULL,
  sw_plot_missing = c(TRUE, FALSE)[1],
  sw_codebook = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{fn_list}{list of files to read}

\item{path_data}{path to data}

\item{path_results_dat}{path to write .RData file}

\item{sw_plot_missing}{T/F plot missing values}

\item{sw_codebook}{T/F generage codebook (only runs if \code{sw_plot_missing} is also \code{TRUE}}
}
\value{
dat_provider_RORA
}
\description{
Read Provider RORA
}
\examples{
\dontrun{
dd_read_provider_RORA(
    fn_list           = c(
                          "FY23 RORA Monitoring Data_20230914.xlsx"
                        )
  , path_data         = "../Data_in/Provider_RORA"
  , path_results_dat  = path_results_dat
  , sw_plot_missing   = c(TRUE, FALSE)[1]
  , sw_codebook       = c(TRUE, FALSE)[1]
  )
}
}
