% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmdohddrisk_logo.R
\name{print.nmdohddrisk_logo}
\alias{print.nmdohddrisk_logo}
\title{Print for nmdohddrisk_logo}
\usage{
\method{print}{nmdohddrisk_logo}(x, ...)
}
\arguments{
\item{x}{the \code{nmdohddrisk_logo()}}

\item{...}{additional parameters passed to \code{cat()}}
}
\value{
the logo, invisibly
}
\description{
Print for nmdohddrisk_logo
}
