% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_features_client_GER.R
\name{dd_features_client_GER}
\alias{dd_features_client_GER}
\title{Features for Client GER}
\usage{
dd_features_client_GER(
  dat_client_GER = NULL,
  dat_client_Match = dat_client_Match
)
}
\arguments{
\item{dat_client_GER}{dat_client_GER data from \code{dd_read_client_GER()}}

\item{dat_client_Match}{dat_client_Match data from \code{dd_read_client_Match()}}
}
\value{
dat_client_GER
}
\description{
Features for Client GER
}
\examples{
\dontrun{
dd_features_client_GER(
    dat_client_GER
  )
}
}
