% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_prediction_model.R
\name{dd_prediction_model}
\alias{dd_prediction_model}
\title{NM DOH DD Prediction Model, main predict and model function}
\usage{
dd_prediction_model(
  params = params,
  path_list = path_list,
  m_months_GER = "4 months",
  m_months_Syncronys = "3 months",
  m_months_Conduent_Omnicad = "4 months",
  m_months_RORA = "4 months",
  m_months_BBS = "2 months",
  m_months_CaseNotes = "4 months",
  sw_rfsrc_ntree = 500,
  sw_alpha = 0.2,
  sw_unit_of_analysis = c("Client_System_ID", "Client_System_ID__ANE_Date")[1],
  sw_select_full = c("select", "full")[1],
  var_subgroup_analysis = "Client_Waiver",
  sw_imbalanced_binary = c(FALSE, TRUE)[1],
  sw_quick_full_only = c(FALSE, TRUE)[1],
  sw_m_months_select_quick_full = FALSE,
  sw_m_months_values = list(GER = c(1, 2), Syncronys = c(1, 2), Conduent_Omnicad = c(1),
    RORA = c(1), BBS = c(1), CaseNotes = c(1))
)
}
\arguments{
\item{m_months_GER}{GER       data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_Syncronys}{Syncronys data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_Conduent_Omnicad}{Omnicad   data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_RORA}{RORA      data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_BBS}{BBS       data, date range to keep, from first ANE or last observation back number of months}

\item{sw_rfsrc_ntree}{passed to \code{e_rfstr_classification()}, Random forest, number of trees}

\item{sw_alpha}{passed to \code{e_rfstr_classification()}, Random forest, model selection alpha level}

\item{sw_unit_of_analysis}{ID for first ANE or (ID, Date) for all ANE.}

\item{sw_select_full}{passed to \code{e_rfstr_classification()}, run RF with model selection, or only fit full model}

\item{var_subgroup_analysis}{passed to \code{e_rfstr_classification()}, variable(s) list (in \code{c(var1, var2)}) for subgroup analysis (group-specific ROC curves and confusion matrices) using ROC threshold from non-subgroup ROC curve, or \code{NULL} for none}

\item{sw_imbalanced_binary}{passed to \code{e_rfstr_classification()}, T/F to use standard or imbalanced binary classification with \code{rfsrc::imbalanced()}.  Also increases ntree to \code{5 * sw_rfsrc_ntree}.}

\item{sw_quick_full_only}{passed to \code{e_rfstr_classification()}, T/F to only fit full model and return model object}

\item{sw_m_months_select_quick_full}{T/F to only search for best \code{m_months} arguments.}

\item{sw_m_months_values}{if \code{sw_m_months_select_quick_full == TRUE}, list of number of months for each data source to perform model selection on m_months arguments.  One strategy for search is to start with \code{c(1, 2)} and if the result is 1, then keep 1, otherwise try \code{c(2, 3)}, and so on for 4, 6, 9, and 12 months.  Increment only 4 values at a time for 16 processors.}

\item{working_directory}{Working directory}

\item{path_prefix_out}{Path and file prefix for output folders}

\item{DD_Group}{All or specific Waiver group}

\item{date_Current}{Date for last day of data to use for analysis}
}
\value{
dat_client_RORA
}
\description{
NM DOH DD Prediction Model, main predict and model function
}
\examples{
\dontrun{
dd_read_client_RORA(
    fn_list           = c(
                          "FY23 RORA Monitoring Data_20230914.xlsx"
                        )
  , path_data         = "../Data_in/Client_RORA"
  , path_results_dat  = path_results_dat
  , sw_plot_missing   = c(TRUE, FALSE)[1]
  , sw_codebook       = c(TRUE, FALSE)[1]
  )
}
}
