% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_prediction_model.R
\name{dd_prediction_model}
\alias{dd_prediction_model}
\title{NM DOH DD Prediction Model, main predict and model function}
\usage{
dd_prediction_model(
  working_directory =
    "D:/Dropbox/StatAcumen/consult/Industry/2023_NM_DOH_DDWaiver/data/Zdrive/Risk_Prediction_Model/Scripts",
  path_prefix_out = "out_V06",
  DD_Group = c("All", "DD", "MV", "MF")[1],
  date_Current = "2023-09-06",
  m_months_GER = "6 months",
  m_months_Syncronys = "2 months",
  m_months_Conduent_Omnicad = "2 months",
  m_months_RORA = "2 months",
  m_months_BBS = "12 months",
  m_months_CaseNotes = "2 months",
  sw_rfsrc_ntree = 500,
  sw_alpha = 0.2
)
}
\arguments{
\item{working_directory}{Working directory}

\item{path_prefix_out}{Path and file prefix for output folders}

\item{DD_Group}{All or specific Waiver group}

\item{date_Current}{Date for last day of data to use for analysis}

\item{m_months_GER}{GER       data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_Syncronys}{Syncronys data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_Conduent_Omnicad}{Omnicad   data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_RORA}{RORA      data, date range to keep, from first ANE or last observation back number of months}

\item{m_months_BBS}{BBS       data, date range to keep, from first ANE or last observation back number of months}

\item{sw_rfsrc_ntree}{Random forest, number of trees}

\item{sw_alpha}{Random forest, model selection alpha level}
}
\value{
dat_client_RORA
}
\description{
NM DOH DD Prediction Model, main predict and model function
}
\examples{
\dontrun{
dd_read_client_RORA(
    fn_list           = c(
                          "FY23 RORA Monitoring Data_20230914.xlsx"
                        )
  , path_data         = "../Data_in/Client_RORA"
  , path_results_dat  = path_results_dat
  , sw_plot_missing   = c(TRUE, FALSE)[1]
  , sw_codebook       = c(TRUE, FALSE)[1]
  )
}
}
